Refactor apt-get actions in Dockerfiles
Running apt-get in a Docker image is a heavy operation. The process that
Docker uses to save and reuse 'layers' between container images really
necessitates that `apt-get update` is used in each RUN where `apt-get
install` is called. Restructure the existing Docker logic to broadly
follow the guidelines from the upstream project:
<https://docs.docker.com/engine/userguide/eng-image/dockerfile_best-practices/>

Change-Id: Ia8f94129cd2792700a8649f6c813b110eff16f32