Add helpers for local testing
Some of us are horrible people who do not run Linux as a local
development environment. The Vagrantfile attempts to build a vm with
support for running build.py and docker locally. It is a very quick and
dirty proof of concept at this point, but seems to work.

The run-image script makes testing an image easier as well. It is
especially nice when using Vagrant:

  vagrant up --provision
  vagrant ssh -- toollabs-images/build.py python-web
  vagrant ssh -tc 'toollabs-images/run-image python-web'

Change-Id: I47ae0d38b987e5a18908d922390723e632af5572