sssd: Add some new images to test sssd in containers
Since we are rather tightly interwoven with LDAP from the webservice layer up,
we need names to resolve in containers.  Since nslcd has failed us in some
OS versions, prompting us to move to sssd, we need to get that working in containers as well.

This is a very simple mirror of the technique used for nslcd except that it
intends to use the minimum sss userland and to mount an entire directory of sockets
instead of a direct socket mount in the container.  I'm hoping this will
be less fragile than the socket.

Bug: T229058
Change-Id: I87ff75fb6de08486f9e84ce31f87d449a69a361b