jessie: Work around removal of jessie-backports
Apply a series of config and runtime changes to make the Jessie
container derived images buildable again following the removal of the
jessie-backports apt repo. We will be working to phase out use of Jessie
images entirely, but that will take a few months so we need to be able
to keep patching/upgrading these in the meantime.

Change-Id: Ie257f596ca85d94ad0a35c6f0aaee886fd8205a1