Add black to tox
Add a tox environment that will run black in diff mode to show
formatting changes that should be applied to the codebase.
A pyproject.toml configuration file is also added to codify the tweaks
to black's defaults which the WMCS team is currently using.

The existing tests/build_tests.py script is also changed here by
applying the diff suggested by black.

Change-Id: I08da49ba9f53c4e7211505d84455fc347f6d7a2a