Introduce jinja2 templating
We are doing far too much cut-n-paste duplication in our
Dockerfile.templates. Begin to change this by adding a runtime
dependency on the Jinja2 template library. This first iteration does the
most basic substitution of prior behavior. Actually using some of the
new features this introduces will be done in followup patches.

Change-Id: If329b32aba9ad4e5dead93c022db10997a715f20