rails: ruby needs nodejs and bundler to be useful
For anybody to run basically any ruby program, they'll want bundler at
a minimum to install the gems they need. Also, rails will not work
without nodejs and yarn at this point. Removing system-installed unicorn because
I don't think that will be useful or performant.

This is the minimum viable version to run bundle install for a rails app
and actually be able to get to a point where it starts.

Bug: T141388
Bug: T250118
Change-Id: If82306bc33a878b9f8ecf874c0972ecfc5ad2736