jessie: try to make jessie containers build at least one more time
Right now, the archive.debian.org repo's GPG keys is marked expired.
Apparently apt_options="--force-yes" should fix that. Any place where
packages might be installed from there, I'm aiming to add that option.

This is linked to the below bug only in the sense that rebuilding
the containers is essential to deploying that code.

Bug: T197930
Change-Id: Ic90f93a805dd084f8d8a183bbf7d2d5025532cbc