Add buildpack images ("stacks")
The build image is the base layer when the buildpack is being built.
Everything runs as the `tfb` user, which has a fixed UID/GID. It
includes `build-essential` packages to provide gcc, make, etc. that are
really only ever needed at build time.

The install-packages script allows for installing Debian packages
without being root. It's super hacky but does the job for now. It's
lightly forked from Heroku's apt buildpack, which is MIT licensed.

The run image is the base layer when the buildpack image is being run in
k8s. Again, everything runs as the same `tfb` user. It has no extra
dependencies on top of the base image.

While this does duplicate some of the buster-sssd image, it should be
clear that these images are a new generation of images, separate from
the webservice ones.

These images should be compliant with the spec:
<https://github.com/buildpacks/spec/blob/main/platform.md#stacks>.

Bug: T265686
Change-Id: I427e2f52d17299e3fb8ca2ec964bc9bb517c84f2