static-binaries: first pass at a stripped-down image for binaries
With growing interest in golang and rust in the Foundation and the community,
it seems sensible to have a more appropriate image to run things like that
in. This produces an image that is much less than half the size of the
golang image (which confusingly seems to be a compiler rather than a
runner). Since all images are now sssd, except buildpacks, I've removed
the sssd from the name of the new image definition.

This is a first pass which should be suitable for job running before
making changes to tools-webservice

Bug: T277749
Change-Id: I60a82e58515d8b757e3f8c041ce82f33f1787f6e