Add yarn to node images
Yarn is a package manager commonly used in the
JavaScript ecosystem.
A package that I plan on deploying as a tool has
a dependency on yarn, and rather than install it
within my project it might be best made accessible
to all users of the node images on tool forge.

This will install yarn 1.22.x currently.

This is works with all of these version of node:
Node: ^4.8.0 || ^5.7.0 || ^6.2.2 || >=8.0.0
per https://classic.yarnpkg.com/en/docs/install

The approach was copied from 2 ruby images that
currently install yarn.

At some point we might want to update to yarn 2.
But that doesn't seem needed right now.

Change-Id: Ibc1f26bf74587395de863ce7aad5210933c14e24