python39: Use shell reimplementation of webservice-runner
As a test for T293552, replace the Python `webservice-runner` with a
reimplementation in a simpleish shell script. This is a direct port of
toolsws.wstypes.python with the port hardcoded for Kubernetes.

`webservice` and existing pods will continue to pass arguments to
`webservice-runner` however those will all be ignored as this runner is
already Python-specific.

Because this is already the Python runtime, removing the
toolforge-webservice package should have a minimal impact. However, the
following packages will no longer be installed in the container:
  python3-certifi python3-chardet python3-idna python3-requests
  python3-urllib3 python3-yaml

In the super rare case that someone was expecting those but didn't set up a
venv for a framework, they can just install them with a venv.

Also take the opportunity to stop loading the uwsgi "python" plugin that
doesn't exist in Python 3 and emits a confusing warning on startup.

Bug: T293552
Change-Id: I49a35bb950221a091f6a04bd46a66bbf529281f8