Use shell webservice-runner for node16 image
See 9db8bc6b989 for details/rationale on this change. Notably
we explicitly install Python3 in this container just in case
someone was actually relying on it being in the container.

The shell script differs from the packaged `tools-webservice`
implementation in that we don't look for `/usr/local/bin/npm`,
which was just used on the grid when we used a integration/npm
repository.

Bug: T293552
Change-Id: Ib809cc777a4ceb44e7270d1641477add2b68fef9