php74: add many TTF fonts
Add a collection of font packages to support rsvg image rendering by the
svgtranslate tool (and hopefully some other tools as well).

The set of fonts installed is not the complete set available on the
Toolforge bastions and grid engine exec nodes. It is only the subset of
packages named 'font-*' which ideally is most TTF/OTF fonts which have
historically been available in those environments.

Bug: T310435
Change-Id: I5a06ce25412974b67ffb9a8923ad69c0f6fda465