Remove jessie-sssd & stretch-sssd image configuration
Follow up to 7fc6331 where we removed all of the children of these
images, but missed removing the base images themselves.

Actually removing the built images from our image registry is a separate
and distinct step that should only be done following an announcement and
migration assistance period for Toolforge tool maintainers.

Change-Id: I8865ba8e49fcef40fbeb21186f69109eb18e98a7