{bullseye,buster}-sssd: add openssh-client
For better or worse, folks have strong opinions about using ssh as a git
transport. Let's keep the world moving by enabling git+ssh use rather
than engaging in endless debate over what is "necessary".

This change adds approximately 5MB to the Bullseye image. The delta for
the Buster image has not been measured but is assumed to be similar.

Bug: T258841
Change-Id: I49920dd7853c385380b2a96a8a9f115aa3a1bee9