Add php82 images
Featuring a shell-based webservice runner (T293552). I'm not a huge fan
of having all of the lighttpd config as it's quite long and will be
copied to other images, but I think that's a better option than any
others we have available today.

Also, compared to the PHP 7.4 images, this image
 * does not install the lighttpd-mod-openssl package just to disable the
   SSL support, and
 * No longer supports serving Python or Perl as CGI scripts.

Bug: T335352
Change-Id: If70ee4d3a3cc7e68a10a5ff2a3638aec497b5542