Fix broken vim modelines
Fix vim modelines that fail in vim 8.2 with "E518: Unknown option: #}".

There are two forms of vim modelines. The first one runs to the end of
the line so is not suitable here. The second one is terminated by a
colon, requires the "set" keyword, and requires options to be separated
by spaces, not colons.

Change-Id: I74323b7f6fa82bc8313bbf7082e9aa1ed3524c2d