Add Trixie images
Initial set of Trixie based images. These are more or less copy-pastes
from Bookworm, except that the font packages have been removed from the
PHP image per my proposal on the task. (This might still change to
restoring the packages in the image, but it's much easier to leave them
out during the discussion period than to remove them after tools are
migrated to the new images.)

Bug: T400255
Change-Id: If3ba45ca15044a01e0f27c6f064bc299e6dd8d9c